backend/requirements.txt
# AI 伴学小程序 - 后端依赖

# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0

# 数据库
sqlalchemy==2.0.25

# 工具
python-dotenv==1.0.0
pydantic==2.5.3

# HTTP 客户端 (AI API 调用)
httpx==0.27.0

# 可选：OSS 备份
# oss2==2.18.0
